% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ansi_aware_handler}
\alias{ansi_aware_handler}
\title{Helper function for printing ANSI in Rmarkdown output
https://blog.djnavarro.net/posts/2021-04-18_pretty-little-clis/}
\usage{
ansi_aware_handler(x, options)
}
\description{
Helper function for printing ANSI in Rmarkdown output
https://blog.djnavarro.net/posts/2021-04-18_pretty-little-clis/
}
