% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{brm_cache}
\alias{brm_cache}
\title{Refit BRMS Models without Recompilation}
\usage{
brm_cache(formula, data, ...)
}
\arguments{
\item{formula}{A multiresponse/multi-input formula of the form
\deqn{
 y1 + y2 + ... ~ x1 + x2 + ..
 }
with which to estimate a BRMS model}

\item{data}{A data.frame containing all the variables in the formula and used
as the basis for estimation.}
}
\value{
models A list of estimated BRMS models. The j^th element contains
y\link{j} ~ x1 + x2 + ...
}
\description{
The most time-consuming part of using BRMS in parallel across many responses
is waiting for compilation to complete. It is more efficient instead to
compile the model once and estimate many models using different datasets.
That is the approach adopted in this function, which speeds up over a naive
loop.
}
