% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{brms_sampler}
\alias{brms_sampler}
\title{Sample from a Bayesian Regression Model}
\usage{
brms_sampler(fits, newdata = NULL, indices = NULL, ...)
}
\arguments{
\item{newdata}{A data.frame containing new inputs from which to sample
responses. If NULL, defaults to the data used to estimate fit.}

\item{indices}{The coordinates of the response from which we want to sample.}
}
\description{
This samples from the posterior predictive for each component in
a multiresponse Bayesian Regression model.
}
