% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_profile.R
\name{check_profile}
\alias{check_profile}
\title{Fail Fast for Poorly Specified Profiles}
\usage{
check_profile(object)
}
\arguments{
\item{object}{An object of class \code{treatment_profile} that we want to check.}
}
\description{
Specifying counterfactuals in general mediation analysis is tricky. We
need to have a hypothetical treatment that applies to the mediator and
outcome separately, even though in reality only one treatment can ever be
assigned to a single treatment. This helper checks that the specification of
mediator and outcome treatments is reasonable.
}
