% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{demo_joy}
\alias{demo_joy}
\title{A Demo Dataset (Random)}
\usage{
demo_joy(n_samples = 100, n_mediators = 5, n_pretreatment = 3)
}
\arguments{
\item{n_samples}{How many random samples to generate?}

\item{n_mediators}{How many random mediators to generate?}

\item{n_pretreatment}{How many random pretreatment variables?}
}
\value{
SE The summarized experiment containing random data.
}
\description{
This is a simple dataset with random data, used simply to illustrate the
design of multimedia. There is no real association between any treatments,
mediators, or outcomes. It always returns a single outcome (PHQ)
SummarizedExperiment, and it randomly assigns samples to Treatment and
Control (see the colData). It stores all the hypothetical mediator data in
the assay slot.
}
\examples{
demo_joy()

demo_joy(n_samples = 2, n_mediators = 20)
}
