% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{demo_spline}
\alias{demo_spline}
\title{A Demo Dataset (Spline)}
\usage{
demo_spline(n_samples = 5000, tau = c(2, 2))
}
\arguments{
\item{n_samples}{The number of samples to generate in the toy example}

\item{tau}{The true direct effects associated with the two outcomes. Defaults
to 2, 2.}
}
\value{
xy A data.frame whose columns include the treatment, mediation, and
outcome variables.
}
\description{
This is a simple dataset with nonlinear relationships between the outcome and
mediators. It is used simply to illustrate the design of multimedia. The
mediator->outcome effect is generated from a random spline function.
}
\examples{
demo_spline()
}
