% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{direct_effect}
\alias{direct_effect}
\title{Direct Effects from Estimated Model}
\usage{
direct_effect(model, t1 = NULL, t2 = NULL, t_mediator = NULL)
}
\description{
Direct Effects from Estimated Model
}
