% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{edges,multimedia-method}
\alias{edges,multimedia-method}
\title{Access Mediation Model DAG}
\usage{
\S4method{edges}{multimedia}(object)
}
\arguments{
\item{object}{An object of class multimedia.}
}
\value{
A data.frame whose rows give edges in the mediation analysis DAG.
}
\description{
This is an accessor to the edges slot in a multimedia object. It is the
internal representation of the variable conditional dependence graph encoded
by the mediation model's DAG.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
multimedia(exper) |>
    edges()
}
