% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{edges}
\alias{edges}
\title{Graphical Structure for Mediation Objects}
\usage{
edges(object)
}
\description{
We often want to access the DAG for different mediation-related S4 objects.
This generic helps us access these graphical model edges lists. See method
instantiations for specific examples.
}
