% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{edges_df}
\alias{edges_df}
\title{Define Edges associated with the Mediation Analysis DAG}
\usage{
edges_df(edges)
}
\arguments{
\item{edges}{A tidygraph graph storing edges between all variables in the
mediation analysis. For example, if a treatment node leads into a mediation
node, this will be included in the graph (as will all pretreatment to mediator,
mediator to outcome, etc.)}
}
\description{
Given a collection of nodes, define the edges between all possible inputs and
outputs.
}
