% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{effect_summary}
\alias{effect_summary}
\title{Average Effects across j}
\usage{
effect_summary(effects, N = 10)
}
\arguments{
\item{effects}{The output from direct_effect or indirect_effect. A data.frame
containing effect estimates for each variable and indirect/direct setting
along rows.}
}
\value{
A version of the input with all indirect/direct settings averaged
over.
}
\description{
This averages direct or indirect effects across settings j, leading to
the effect estimates given in equation (10) of the preprint.
}
\seealso{
direct_effect indirect_effect
}
