% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimator}
\alias{estimator}
\title{Accessor for Model Estimators}
\usage{
estimator(object)
}
\arguments{
\item{object}{An object of class \code{model} whose estimator we would like to
call.}
}
\value{
A function that can be called with formula and data arguments, like
}
\description{
Accessor for Model Estimators
}
\examples{
m <- lm_model()
estimator(m)(mpg ~ hp + wt, data = mtcars)
}
