% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{expand_edges}
\alias{expand_edges}
\title{Helper for \code{graph_specification}}
\usage{
expand_edges(input, output, input_name, output_name)
}
\arguments{
\item{input}{A vector of names for all variables within the input group.}

\item{output}{A vector of names for all variables within the output group.}

\item{input_name}{A node_type to associate with all inputs. For example,
"treatment" or "mediator" for edges along the T->Y or X->M paths.}

\item{output_name}{A node_type to associate with all outputs. For example,
"mediator" or "outcome" for edges along the T->M or M->Y paths.}
}
\value{
G A tidygraph graph with bipartite structure. One set of nodes
corresponds to the inputs, one to the outputs, and every pair of inputs and
outputs are linked.
}
\description{
To define the overall mediation analysis graph, it helps to have a helper
that learns all combinations of variables between two node types. This is
done simply by using expand.grid and converting into a bipartite tidygraph.
}
