% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{exper_df}
\alias{exper_df}
\title{Convert a Summarized Experiment to a data.frame}
\usage{
exper_df(exper)
}
\arguments{
\item{exper}{An object of class SummarizedExperiment.}
}
\description{
Convert a Summarized Experiment to a data.frame
}
