% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fdr_summary}
\alias{fdr_summary}
\title{Calibration using Synthetic Nulls}
\usage{
fdr_summary(contrast, effect = "indirect_overall", q_value = 0.15)
}
\arguments{
\item{effect}{Either "indirect_overall" (the default), "indirect_pathwise",
or "direct_effect" specifying the type of effect that we want to select.}

\item{q_value}{The target for false discovery rate control. The last time the
estimated FDR is above this threshold is smallest magnitude of effect size
that we will consider.}
}
\value{
fdr A data.frame specifying, for each candidate effect, whether it
should be selected.
}
\description{
This function computes a threshold for indirect or direct effect estimates
that controls the false discovery rate according to estimates made using real
and synthetic null data, against the null hypotheses that effects are zero.
It computes the proportion of synthetic null estimates that are among the top
K largest effects (in magnitude) as an estimate of the FDR.
}
