% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{from_summarized_experiment}
\alias{from_summarized_experiment}
\title{Helper to Convert SummarizedExperiment to a data.frame}
\usage{
from_summarized_experiment(
  exper,
  outcomes,
  treatments,
  mediators,
  pretreatments
)
}
\arguments{
\item{outcomes}{A vector containing names of all outcome variables.}

\item{treatments}{A vector containing names of all treatment variables.}

\item{mediators}{A vector containing names of all mediators.}

\item{pretreatments}{A vector containing names of all pretreatment variables.
Defaults to NULL, in which case the pretreatments slot will remain empty.}

\item{x}{A SummarizedExperiment whose colData or assays contains all the
variables needed for the mediation analysis.}
}
\value{
result An object of class \code{mediation_data}, with separate slots for
each of the node types in a mediation analysis diagram.
}
\description{
Helper to Convert SummarizedExperiment to a data.frame
}
