% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{glmnet_model}
\alias{glmnet_model}
\title{Regularized Glmnet Model across Responses}
\usage{
glmnet_model(...)
}
\arguments{
\item{...}{Keyword parameters passed to glmnet.}
}
\value{
model An object of class \code{model} with estimator, predictor, and
sampler functions associated wtih a lienar model.
}
\description{
Apply a regularized (generalized) linear model in parallel across each
response $y$ in an outcome or mediation model. This can be helpful when we
have many mediators or pretreatment variables, making the input
high-dimensional.
}
\seealso{
model lm_model rf_model
}
