% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{glmnet_sampler}
\alias{glmnet_sampler}
\title{Sample from a Glmnet Model}
\usage{
glmnet_sampler(fits, newdata = NULL, indices = NULL, lambda_ix = 1, ...)
}
\description{
This assumes a continuous response, so that the out-of-sample MSE can be used
to estimate the outcome variability $\sigma$.
}
