% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{graph_specification}
\alias{graph_specification}
\title{Define all Edges in a Mediation Analysis}
\usage{
graph_specification(outcomes, treatments, mediators, pretreatments)
}
\arguments{
\item{outcomes}{A vector containing names of all outcome variables.}

\item{treatments}{A vector containing names of all treatment variables.}

\item{mediators}{A vector containing names of all mediators.}

\item{pretreatments}{A vector containing names of all pretreatment variables.}
}
\value{
edges A tidygraph graph containing all nodes and edges in the
mediation analysis. Each node is annotated with a \code{node_type}, describing
which of the parts of the mediation analysis graph the node belongs to.
}
\description{
This processes the variable names from a mediation analysis and builds a
graph representing causal relationships in the mediation analysis DAG. It is
helpful for automatically building formulas for the mediation and outcome
models, and it is used to create the \code{edges} slot in multimedia objects.
}
\seealso{
multimedia
}
