% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{indirect_pathwise}
\alias{indirect_pathwise}
\title{Indirect Effects via Single Mediation Paths}
\usage{
indirect_pathwise(model, t1 = NULL, t2 = NULL, t_outcome = NULL)
}
\description{
Indirect Effects via Single Mediation Paths
}
\examples{
model <- multimedia(
  demo_joy(),
  outcomes = "PHQ",
  treatments = "treatment",
  mediators = starts_with("ASV")
) |> estimate(demo_joy())
t1 <- tibble(treatment = as.factor("Treatment"))
t2 <- tibble(treatment = as.factor("Control"))
t_outcome <- list(t1, t2)
indirect_pathwise(model, t1, t2, t_outcome)
indirect_pathwise(model, "Treatment", "Control")
}
