% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{matching_indices}
\alias{matching_indices}
\title{Parse an Edgelist for Mediation or Outcome Edges}
\usage{
matching_indices(edges, nulls = NULL)
}
\arguments{
\item{edges}{The full edgelist associated with the current mediation analysis
(mediation + outcome) models.}

\item{nulls}{A string specifying the indices of edges to ignore. "T->Y",
"T->M", and "M->Y" will match all edges between treatment to outcome,
treatment to mediator, etc. Otherwise, the vector of indices specifying
which edges to ignore.}
}
\value{
The ID of edges that match the \code{nulls} parameter.
}
\description{
Parse an Edgelist for Mediation or Outcome Edges
}
