% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\docType{class}
\name{mediation_data-class}
\alias{mediation_data-class}
\title{An S4 Class Representing a Mediation Analysis Dataset}
\description{
All the data needed for carrying out mediation analysis, encapsulated in one
object. Each node in the mediation analysis DAG is represented by a
data.frame in its own slot. Pretreatments are optional and will be left NULL
if not specified.
}
\section{Slots}{

\describe{
\item{\code{mediators}}{A data.frame containing observed values of the mediators
across all samples.}

\item{\code{outcomes}}{A data.frame containing observed values of the outcomes across
all samples.}

\item{\code{treatments}}{A data.frame containing observed values of the treatments
across all samples.}

\item{\code{pretreatments}}{A data.frame containing observed values of the
pretreatments across all samples.}
}}

