% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{mediation_models}
\alias{mediation_models}
\title{Accessor for Outcome Models}
\usage{
mediation_models(object)
}
\arguments{
\item{object}{An object of class multimedia whose outcome model estimates we
would like to extract.}
}
\value{
A list containing all the fitted mediation models.
}
\description{
Accessor for Outcome Models
}
\examples{
data(mindfulness)
exper <- mediation_data(
    mindfulness,
    phyloseq::taxa_names(mindfulness),
    "treatment",
    starts_with("mediator"),
    "subject"
)

m <- multimedia(exper)
mediation_models(m)
}
