% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{mediators,mediation_data-method}
\alias{mediators,mediation_data-method}
\title{Access to @mediators in Mediation Data}
\usage{
\S4method{mediators}{mediation_data}(object)
}
\arguments{
\item{object}{An object of class mediation_data.}
}
\value{
m A data.frame whose rows are samples and columns are values of
mediators across those samples.
}
\description{
This is an accessor returns the @mediators slot in a mediation_data object.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
mediators(exper)
}
