% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{mediators<-,mediation_data-method}
\alias{mediators<-,mediation_data-method}
\title{Set the Mediators in a Mediation Data Object}
\usage{
\S4method{mediators}{mediation_data}(object) <- value
}
\arguments{
\item{object}{An object of class mediation_data.}

\item{value}{The new mediators values for the object.}
}
\value{
A version of \code{object} whose mediators slot has been replaced.
}
\description{
This is an setter method for the mediators slot in a mediation data object.
It lets you supply a new mediators data.frame for the object.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
mediators(exper) <- data.frame(m = 1:10)
exper
}
