% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{mediators<-}
\alias{mediators<-}
\title{Set Mediators}
\usage{
mediators(object) <- value
}
\arguments{
\item{object}{An object whose mediators slot to modify.}

\item{values}{The new mediator values to set within object.}
}
\description{
This is an setter method for mediators in an S4 objet, usually of class
mediation_data.
}
