% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{mediators}
\alias{mediators}
\title{Access Class-Specific Mediators}
\usage{
mediators(object)
}
\arguments{
\item{object}{An object whose mediators we want to access (either their names
or values).}
}
\value{
The output will depend on the class of the object that is passed in.
For multimedia objects, this returns a character vector of mediators. For
mediation data objects, this returns the mediator mediator data.
}
\description{
This is a shorthand for accessing mediator-related slots in classes exported
by the multimedia package.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
mediators(exper)
}
