% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{mindfulness}
\alias{mindfulness}
\title{Mindfulness Dataset}
\description{
Data from a study of the relationship between mindfulness and the microbiome,
stored as a phyloseq object. Measurements are from 50 subjects before and
after a real and active control mindfulness intervention. We are interested
in changes in subject-level metadata, stored in the sample_data slot
}
\keyword{data}
