% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\docType{class}
\name{multimedia-class}
\alias{multimedia-class}
\title{An S4 Class Representing Mediation Analysis Models}
\description{
This is the primary class exported by the multimedia package. It includes
both the outcome and mediation models. It also stores complementary data
helpful for estimating effects and carrying out inference. Most estimation
and inference functions expect an input model of this form.
}
\section{Slots}{

\describe{
\item{\code{edges}}{A tidygraph graph specifying all edges in the mediation analysis
DAG. For example, it should include all edges between treatments and
outcomes if we hypothesize the presence of a direct effect.}

\item{\code{outcome}}{A \code{model} class storing the outcome model.}

\item{\code{mediation}}{A \code{model} class storing the mediation model.}

\item{\code{treatments}}{A \code{treatment_profile} class describing treatment assignments
across all samples.}
}}

