% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{n_mediators}
\alias{n_mediators}
\title{Number of Mediators in a Multimedia Object}
\usage{
n_mediators(object)
}
\arguments{
\item{object}{An object of class multimedia.}
}
\value{
An integer specifying the number of mediators.
}
\description{
Number of Mediators in a Multimedia Object
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
multimedia(exper) |>
    n_mediators()
}
