% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{n_outcomes}
\alias{n_outcomes}
\title{Number of Outcomes in a Multimedia Object}
\usage{
n_outcomes(object)
}
\arguments{
\item{object}{An object of class multimedia.}
}
\value{
An integer specifying the number of outcomes.
}
\description{
Number of Outcomes in a Multimedia Object
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
multimedia(exper) |>
    n_outcomes()
}
