% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{outcome_formula}
\alias{outcome_formula}
\title{Generate an Outcome Formula}
\usage{
outcome_formula(edges)
}
\arguments{
\item{edges}{A tidygraph graph storing edges between all variables in the
mediation analysis. For example, if a treatment node leads into a mediation
node, this will be included in the graph (as will all pretreatment to
mediator, mediator to outcome, etc.)}
}
\value{
A formula object with outcomes on the LHS.
}
\description{
Given the relationship between all variables in the mediation analysis DAG,
return a formula of the form
}
\details{
m1 + m2 + ... ~ treatment_1 + ... + pretreatment_T

where all nodes leading into the mediators are inputs and all the mediators
are the responses.
}
