% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{outcome_model}
\alias{outcome_model}
\title{Access the Outcome Model in a Multimedia Object}
\usage{
outcome_model(object)
}
\arguments{
\item{object}{An object of class multimedia}
}
\description{
This is an accessor to the outcome slot of a multimedia object.
}
\examples{
exper <- demo_joy() |>
  mediation_data("PHQ", "treatment", starts_with("ASV"))
multimedia(exper) |>
  outcome_model()
}
