% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{outcomes<-}
\alias{outcomes<-}
\title{Set Outcomes
This is an setter method for outcomes in an S4 object, usually of class
mediation_data.}
\usage{
outcomes(object) <- value
}
\arguments{
\item{object}{An object whose outcomes slot to modify.}

\item{value}{The new outcome values to set within object.}
}
\value{
Modifies the outcomes slot of the input object in place.
}
\description{
Set Outcomes
This is an setter method for outcomes in an S4 object, usually of class
mediation_data.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
outcomes(exper) <- data.frame(new_PHQ = rnorm(100))
}
