% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{outcomes}
\alias{outcomes}
\title{Set Outcomes}
\usage{
outcomes(object)
}
\arguments{
\item{object}{An object whose outcomes slot to modify.}

\item{values}{The new outcome values to set within object.}
}
\description{
This is an setter method for outcomes in an S4 objet, usually of class
mediation_data.
}
