% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict,multimedia-method}
\alias{predict,multimedia-method}
\title{Predictions from a Multimedia Class}
\usage{
\S4method{predict}{multimedia}(object, profile = NULL, mediators = NULL, pretreatment = NULL, ...)
}
\description{
This generalizes the built-in predict method to the multimedia class.  Given
an estimated multimedia object, this function supports prediction along the
estimated DAG. It first predicts \verb{hat[M] | T, X} and then \verb{hat[Y] | hat[M], T, X}. Each prediction step will call the prediction method within the
mediation and outcome models that make up the multimedia object on which this
is called.
}
