% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_across}
\alias{predict_across}
\title{Predictions from a Single Outcome}
\usage{
predict_across(object, newdata, name)
}
\arguments{
\item{object}{An object of class \code{model} containing an estimated model.}

\item{newdata}{A data.frame containing new inputs from which to sample
responses. If NULL, defaults to the data used to estimate fit.}

\item{name}{A string or index specifying which of the dimensions of a
multiresponse prediction to extract.}
}
\description{
Predictions from a Single Outcome
}
