% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{pretreatments<-,mediation_data-method}
\alias{pretreatments<-,mediation_data-method}
\title{Set the Pretreatments in a Mediation Data Object}
\usage{
\S4method{pretreatments}{mediation_data}(object) <- value
}
\arguments{
\item{object}{An object of class mediation_data.}

\item{value}{The new pretreatment values for the object.}
}
\value{
A version of \code{object} whose pretreatments slot has been replaced.
}
\description{
This is an setter method for the pretreatments slot in a mediation data
object. It lets you supply a new pretreatments data.frame for the object.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
pretreatments(exper) <- data.frame(x = 1:10)
exper
}
