% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{sample,multimedia-method}
\alias{sample,multimedia-method}
\title{Sample New Mediator/Outcome Data}
\usage{
\S4method{sample}{multimedia}(x, size, pretreatment = NULL, profile = NULL, mediators = NULL, ...)
}
\description{
This generalizes the built-in sample method to the multimedia class. Given an
estimated multimedia object, this function supports sampling along the
estimated DAG. It first samples \verb{M* | T, X} and then \verb{Y* | M*, T, X}. Each
sampling step will call the sample method within the mediation and outcome
models that make up the multimedia object on which this is called.
}
\examples{
exper <- demo_spline(tau = c(2, 1)) |>
  mediation_data(starts_with("outcome"), "treatment", "mediator")
fit <- multimedia(exper) |>
  estimate(exper)
samples <- sample(fit)
samples@mediators
samples@outcomes

# sampling with just different "n" has no effect.
samples <- sample(fit, 100)

# Instead sample at a new treatment configuration
t1 <- data.frame(treatment = factor(rep(c(0, 1), each = 50)))
profile <- setup_profile(fit, t_mediator = t1, t_outcome = t1)
samples <- sample(fit, profile = profile)
samples@mediators
samples@outcomes
}
