% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity Analysis for Overall Indirect Effect}
\usage{
sensitivity(model, exper, n_rho = 10, n_bootstrap = 100, progress = TRUE)
}
\description{
Sensitivity Analysis for Overall Indirect Effect
}
