% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity_pathwise}
\alias{sensitivity_pathwise}
\title{Sensitivity Analysis for Pathwise Indirect Effects}
\usage{
sensitivity_pathwise(
  model,
  exper,
  n_rho = 10,
  n_bootstrap = 100,
  progress = TRUE
)
}
\description{
Sensitivity Analysis for Pathwise Indirect Effects
}
