% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity_sample}
\alias{sensitivity_sample}
\title{Sample from an SEM with Correlated Errors}
\usage{
sensitivity_sample(model, exper, rho)
}
\description{
Sample from an SEM with Correlated Errors
}
\examples{
xy_data <- demo_spline()
exper <- mediation_data(xy_data, starts_with("outcome"), "treatment", "mediator")
model <- multimedia(exper, outcome_estimator = rf_model(num.trees = 1e3)) |>
 estimate(exper)
sensitivity_sample(model, exper)
}
