% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_profile.R
\name{setup_profile}
\alias{setup_profile}
\title{Define a \code{treatment_profile} object}
\usage{
setup_profile(x, t_mediator = NULL, t_outcome = NULL)
}
\arguments{
\item{x}{An object of class \code{mediation_data} with a slot \verb{@treatments}
containing information on the treatments that have been applied.}

\item{t_mediator}{A data.frame whose columns store treatment names and whose
values are the treatment assignments to each sample (row). Defaults to
NULL, in which case this type of data.frame is constructed from the
treatment assignments in the \code{mediation_data}'s \verb{@treatment} slot. Each
column must be either a numeric or factor variable.}

\item{t_outcome}{A data.frame analogous to \code{t_mediator}, but applying to the
outcome node.}
}
\value{
An object of class \code{treatment_profile} giving treatment assignments
for both mediation and outcome terms.
}
\description{
For general mediation analysis, we need to provide counterfactuals for both
the outcome and mediator components of each sample. That is, we need to
understand Y(t, M(t')) where t and t' may not be the same.
\code{treatment_profile} classes place some more stringent requirements on the
structure of treatment profiles, so that later effect estimation can make
simplifying assumptions. This function creates a treatment profile from a
collection of possible mediator and outcome treatments.
}
\seealso{
check_profile
}
