% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\name{spline_fun}
\alias{spline_fun}
\title{Generate Random Spline}
\usage{
spline_fun(D = 2, knots = NULL, h_ix = 1:10, ...)
}
\description{
Generate Random Spline
}
\examples{
x <- seq(-2, 2, length.out = 100)
f <- spline_fun(sd = 0.3)
fx <- f(x)
plot(x, fx[, 1])
}
