% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{[,mediation_data,ANY,ANY,ANY-method}
\alias{[,mediation_data,ANY,ANY,ANY-method}
\title{Subset a mediation dataset}
\usage{
\S4method{[}{mediation_data,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
We can subset samples by indexing into a mediation dataset. It will subsample
all fields -- pretreatments, treatments, mediators, and outcomes. Note that
there is no way to subset columns in this way, since they would be different
across each source.
}
\examples{
exper <- demo_joy() |>
  mediation_data("PHQ", "treatment", starts_with("ASV"))
exper[1]
exper[1:10]
}
