% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{sub_formula}
\alias{sub_formula}
\title{Helper to Modify Formulas}
\usage{
sub_formula(formula, yj)
}
\arguments{
\item{formula}{The original formula whose response we want to modify.}

\item{yj}{The desired response term for the formula.}
}
\value{
A new formula object with the LHS replaced by yj.
}
\description{
Helper to Modify Formulas
}
