% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_profile.R
\docType{class}
\name{treatment_profile-class}
\alias{treatment_profile-class}
\title{Define a Treatment Profile}
\description{
This class ensures appropriate structure of the treatment assignments for
mediator and outcome variables. It enforces certain structural requirements
(e.g., that the number of samples is the same under the mediator and outcome
counterfactuals) using the \code{check_profile} function.
}
\seealso{
setup_profile check_profile
}
