% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{treatments,mediation_data-method}
\alias{treatments,mediation_data-method}
\title{Treatments in a Mediation Data Object}
\usage{
\S4method{treatments}{mediation_data}(object)
}
\arguments{
\item{object}{An object of class mediation_data.}
}
\value{
A data.frame whose rows are samples and columns different treatments.
}
\description{
This is an accessor function to the @treatments slot in a mediation data
object. It returns the entire set of observed treatments, in contrast to
treatments() applied to a multimedia object, which only returns the names of
the treatment variables.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
treatments(exper)
}
