% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{treatments,multimedia-method}
\alias{treatments,multimedia-method}
\title{Names of Treatments in a Multimedia Object}
\usage{
\S4method{treatments}{multimedia}(object)
}
\arguments{
\item{object}{An object of class multimedia.}
}
\value{
m A vector of strings containing the names of all the treatments.
}
\description{
This is a helper that returns the names of the treatment variables in an
object of class multimedia. It parses the graph in the DAG specifying the
mediation analysis, and it returns all variables of node type treatment.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
multimedia(exper) |>
    treatments()
}
