% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{treatments<-}
\alias{treatments<-}
\title{Set Treatments
This is an setter method for treatments in an S4 object, usually of class
mediation_data.}
\usage{
treatments(object) <- value
}
\arguments{
\item{object}{An object whose treatments slot to modify.}

\item{value}{The new treatment values to set within object.}
}
\value{
Modifies the treatments slot of the input object in place.
}
\description{
Set Treatments
This is an setter method for treatments in an S4 object, usually of class
mediation_data.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
treatments(exper) <- data.frame(new_treatment = rnorm(100))
}
