% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{ansi_aware_handler}
\alias{ansi_aware_handler}
\title{Pretty Printing}
\usage{
ansi_aware_handler(x, options)
}
\arguments{
\item{x}{A character vector potentially including ANSI.}

\item{options}{Unused placeholder argument.}
}
\value{
A string with HTML reformatted to ensure colors appear in printed
code blocks in rmarkdown output.
}
\description{
Helper function for printing ANSI in Rmarkdown output. Use this at the start
of your Rmarkdown files to include colors in the printed object names in the
final compiled output.
}
\details{
Taken from the post at

https://blog.djnavarro.net/posts/2021-04-18_pretty-little-clis/
}
\examples{
knitr::knit_hooks$set(output = ansi_aware_handler)
options(crayon.enabled = TRUE)
}
