% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{effect_summary}
\alias{effect_summary}
\title{Average Effects across j}
\usage{
effect_summary(effects, N = 10)
}
\description{
This averages direct or indirect effects across settings j, leading to
the effect estimates given in equation (10) of the preprint.
}
