% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimator}
\alias{estimator}
\title{Accessor for Model Estimators}
\usage{
estimator(object)

estimator(object)
}
\arguments{
\item{object}{A model object that we want to estimate.}
}
\value{
A function that can be called with formula and data arguments, like

A fitted version of the input model class.
}
\description{
This defines a generic estimator function, which can be applied to different
multimedia model objects. It creates a unified interface to estimating diverse
mediation and outcome model families.
}
\examples{
m <- lm_model()
estimator(m)(mpg ~ hp + wt, data = mtcars)
m <- lnm_model()
mat <- data.frame(matrix(rpois(250, 10), 25, 10))
colnames(mat) <- paste0("y", 1:6)
fit <- estimator(m)(y1 + y2 + y3 + y4 ~ y5 + y6, mat)
}
