% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{exper_df}
\alias{exper_df}
\title{Convert a Summarized Experiment to a data.frame}
\usage{
exper_df(exper)
}
\arguments{
\item{exper}{An object of class SummarizedExperiment.}
}
\value{
A data.frame combining all slots of a multimedia object.
}
\description{
This is a small helepr function to concatenate the colData and assays within
SummarizedExperiment objects into a single data.frame. This is not necessary
for the essential multimedia workflow, it is only exported for potential
independent interest.
}
\examples{
demo_joy() |>
    exper_df()
}
