% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{indirect_overall}
\alias{indirect_overall}
\title{Overall Indirect Effect}
\usage{
indirect_overall(model, t1 = NULL, t2 = NULL, t_outcome = NULL)
}
\description{
Overall Indirect Effect
}
\examples{
t1 <- tibble(treatment = as.factor("Treatment"))
t2 <- tibble(treatment = as.factor("Control"))
t_outcome <- list(t1, t2)
indirect_overall(model, t1, t2, t_outcome)
indirect_overall(model, "Treatment", "Control")
}
