% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{indirect_pathwise}
\alias{indirect_pathwise}
\title{Indirect Effects via Single Mediation Paths}
\usage{
indirect_pathwise(model, exper = NULL, t1 = 1, t2 = 2)
}
\description{
Indirect Effects via Single Mediation Paths
}
\examples{
# example with null data
exper <- demo_joy() |>
  mediation_data("PHQ", "treatment", starts_with("ASV"))
fit <- multimedia(exper) |>
  estimate(exper)

indirect_pathwise(fit)

# example with another dataset
exper <- demo_spline(tau = c(2, 1)) |>
  mediation_data(starts_with("outcome"), "treatment", "mediator")
fit <- multimedia(exper) |>
  estimate(exper)
indirect_pathwise(fit)
}
