% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{mediators<-}
\alias{mediators<-}
\title{Set Mediators}
\usage{
mediators(object) <- value
}
\arguments{
\item{object}{An object whose mediators slot to modify.}

\item{value}{The new mediator values to set within object.}
}
\value{
Modifies the mediators slot of the input object in place.
}
\description{
This is an setter method for mediators in an S4 object, usually of class
mediation_data.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
mediators(exper) <- data.frame(new_ASV = rnorm(100))
mediators(exper)
}
