% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{mindfulness}
\alias{mindfulness}
\title{Mindfulness Dataset}
\format{
An object of class phyloseq.
}
\value{
A phyloseq object with 307 samples and 55 taxa. Samples are described
by 7 variables (potential mediators) in the sample_data slot. There is no
associated phylogenetic tree.
}
\description{
Data from a study of the relationship between mindfulness and the microbiome,
stored as a phyloseq object. Measurements are from 50 subjects before and
after a real and active control mindfulness intervention. We are interested
in changes in subject-level metadata, stored in the sample_data slot
}
\examples{
data(mindfulness)
}
\keyword{data}
