% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{nrow,mediation_data-method}
\alias{nrow,mediation_data-method}
\title{How many samples in the mediation dataset?}
\usage{
\S4method{nrow}{mediation_data}(x)
}
\arguments{
\item{x}{The \code{mediation_data} object whose number of samples we want to
return.}
}
\description{
How many samples in the mediation dataset?
}
