% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{pretreatments<-}
\alias{pretreatments<-}
\title{Set Pretreatments
This is an setter method for pretreatments in an S4 object, usually of class
mediation_data.}
\usage{
pretreatments(object) <- value
}
\arguments{
\item{object}{An object whose pretreatments slot to modify.}

\item{value}{The new pretreatment values to set within object.}
}
\value{
Modifies the pretreatments slot of the input object in place.
}
\description{
Set Pretreatments
This is an setter method for pretreatments in an S4 object, usually of class
mediation_data.
}
\examples{
exper <- demo_joy() |>
    mediation_data("PHQ", "treatment", starts_with("ASV"))
pretreatments(exper) <- data.frame(new_var = rnorm(100))
}
