% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{[,mediation_data,ANY,ANY,ANY-method}
\alias{[,mediation_data,ANY,ANY,ANY-method}
\title{Subset a mediation dataset}
\usage{
\S4method{[}{mediation_data,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
Subset a mediation dataset
}
